added custom directory definition for serializing classes file
