+++ orange_cb_recsys/runnable_instances.py
@@ -3,8 +3,8 @@ import lzma
 import pickle
 import os
 
-from orange_cb_recsys.content_analyzer import FieldConfig, ExogenousConfig, ContentAnalyzer
-from orange_cb_recsys.content_analyzer.config import ContentAnalyzerConfig
+from orange_cb_recsys.content_analyzer.config import ContentAnalyzerConfig, FieldConfig, ExogenousConfig
+from orange_cb_recsys.content_analyzer.content_analyzer_main import ContentAnalyzer
 from orange_cb_recsys.content_analyzer.embeddings.embedding_source import EmbeddingSource
 from orange_cb_recsys.content_analyzer.exogenous_properties_retrieval import ExogenousPropertiesRetrieval
 from orange_cb_recsys.content_analyzer.field_content_production_techniques.embedding_technique.combining_technique import \
@@ -26,7 +26,7 @@ from orange_cb_recsys.recsys.graphs.graph import Graph
 from orange_cb_recsys.recsys.content_based_algorithm.centroid_vector.similarities import Similarity
 from orange_cb_recsys.recsys.recsys import RecSys
 
-from orange_cb_recsys.evaluation import EvalModel
+from orange_cb_recsys.evaluation.eval_model import EvalModel
 from orange_cb_recsys.evaluation.eval_pipeline_modules.methodology import Methodology
 from orange_cb_recsys.evaluation.partitioning_techniques.partitioning import Partitioning
 from orange_cb_recsys.evaluation.metrics.metrics import Metric
@@ -34,7 +34,6 @@ from orange_cb_recsys.evaluation.eval_pipeline_modules.metric_evaluator import M
 from orange_cb_recsys.evaluation.eval_pipeline_modules.partition_module import Split, PartitionModule
 
 from orange_cb_recsys.utils.class_utils import get_all_implemented_classes
-from orange_cb_recsys.utils.const import root_path
 
 """
 List containing all the base classes in the framework
@@ -50,11 +49,12 @@ base_classes: List[Type] = [
 ]
 
 
-def serialize_classes():
+def get_classes():
     """
-    Serializes all the available framework's base classes
+    Function used to create a dictionary containing the framework's classes' names as keys and the classes themselves
+    as values
 
-    To do so, a dictionary is created from the base_classes list and it will be in the following form:
+    A dictionary is created from the base_classes list and it will be in the following form:
 
         {
         "recsys": RecSys,
@@ -70,13 +70,7 @@ def serialize_classes():
     any implemented subclass, said subclasses will be added to the dictionary.
     If a class isn't abstract and has subclasses, both the class itself and the subclasses will be added to the
     dictionary
-
-    The dictionary will then be serialized in a file called 'classes.xz' and, if the file already exists, it will be
-    overwritten
     """
-    if os.path.isfile(os.path.join(root_path, 'orange_cb_recsys/classes.xz')):
-        os.remove(os.path.join(root_path, 'orange_cb_recsys/classes.xz'))
-
     classes_dict = {}
 
     for base_cls in base_classes:
@@ -84,24 +78,26 @@ def serialize_classes():
         for cls in classes:
             classes_dict[cls.__name__.lower()] = cls
 
-    with lzma.open((root_path + 'orange_cb_recsys/classes.xz'), 'wb') as f:
+    return classes_dict
+
+
+def serialize_classes(output_directory: str = "."):
+    """
+    Serializes all the available framework's base classes (retrieving them using the get_classes() function)
+
+    The dictionary will then be serialized in a file called 'classes.xz' and, if the file already exists, it will be
+    overwritten
+
+    Args:
+        output_directory (str): directory where the "classes.xz" file will be saved, if it doesn't exist it will be
+            created
+    """
+    file_path = os.path.join(output_directory, "classes.xz")
+
+    if not os.path.exists(output_directory):
+        os.mkdir(output_directory)
+
+    classes_dict = get_classes()
+
+    with lzma.open(file_path, 'wb') as f:
         pickle.dump(classes_dict, f)
-
-
-def get_classes():
-    """
-    Retrieves the implemented classes dictionary described in the serialize_classes function documentation from the
-    'classes.xz' file. If the file is not found, a FileNotFoundError exception is thrown
-    """
-    try:
-        path = os.path.join(root_path, 'orange_cb_recsys/classes.xz')
-        with lzma.open(path, "rb") as f:
-            classes = pickle.load(f)
-    except FileNotFoundError:
-        raise FileNotFoundError('File containing classes not found')
-
-    return classes
-
-
-if __name__ == "__main__":
-    serialize_classes()
